Optimisations

+++ src/core/ImageManager.h
@@ -5,8 +5,8 @@
 
 class ImageManager {
 public:
-    // Load PNG, JPG, TGA, BMP, PSD etc. into internal float RGBA buffer
-    static bool LoadImageFromFile(const std::string& filepath, std::vector<float>& outPixels, int& outWidth, int& outHeight);
+    // Load PNG, JPG, TGA, BMP, PSD etc. into RGBA8 buffer.
+    static bool LoadImageFromFile(const std::string& filepath, std::vector<uint8_t>& outPixels, int& outWidth, int& outHeight);
 
     // Save internal float RGBA buffer into PNG or JPG, with optional ICC profile
     static bool SaveImageToFile(const std::string& filepath, const std::vector<float>& pixels, int width, int height, const std::string& iccProfilePath = "");
